Move utils that aren't used by YUI modules to init utils.
