+++ jujugui/static/gui/src/test/test_model.js
@@ -1530,6 +1530,43 @@ describe('test_model.js', function() {
 
     });
 
+    describe('service state simplification', function() {
+      var simplifyState;
+
+      before(() => {
+        const units = new models.ServiceUnitList();
+        simplifyState = units._simplifyState;
+      });
+
+      var makeUnit = function(state, relationErrors) {
+        var unit = {agent_state: state};
+        if (relationErrors) {
+          unit.relation_errors = {myrelation: ['service']};
+        }
+        return unit;
+      };
+
+      it('translates service running states correctly', function() {
+        var unit = makeUnit('started');
+        assert.strictEqual('running', simplifyState(unit));
+      });
+
+      it('translates service error states correctly', function() {
+        var states = ['install-error', 'foo-error', '-error', 'error'];
+        states.forEach(function(state) {
+          var unit = makeUnit(state);
+          assert.strictEqual(simplifyState(unit), 'error', state);
+        });
+      });
+
+      it('translates service pending states correctly', function() {
+        var states = ['pending', 'installed', 'waiting', 'stopped'];
+        states.forEach(function(state, index) {
+          var unit = makeUnit(state);
+          assert.strictEqual(simplifyState(unit), states[index], state);
+        });
+      });
+    });
   });
 
   describe('Charm load', function() {
@@ -2512,6 +2549,61 @@ describe('test_model.js', function() {
       });
     });
 
+    describe('_numToLetter', function() {
+      it('converts numbers to letters correctly', function() {
+        // Map of numbers and output to check. This list isn't exhaustive
+        // but checks some important milestones for common issues with this
+        // technique.
+        var mapping = {
+          1: 'a',
+          2: 'b',
+          10: 'j',
+          15: 'o',
+          26: 'z',
+          27: 'aa',
+          28: 'ab',
+          52: 'az',
+          53: 'ba',
+          54: 'bb',
+          703: 'aaa',
+          748: 'abt',
+          1982: 'bxf'
+        };
+        Object.keys(mapping).forEach(function(key) {
+          assert.equal(
+            list._numToLetter(key), mapping[key],
+            key + ' did not properly convert to ' + mapping[key]);
+        });
+      });
+    });
+
+    describe('_letterToNum', function() {
+      it('converts letters to numbers correctly', function() {
+        // Map of numbers and output to check. This list isn't exhaustive
+        // but checks some important milestones for common issues with this
+        // technique.
+        var mapping = {
+          a: 1,
+          b: 2,
+          j: 10,
+          o: 15,
+          z: 26,
+          aa: 27,
+          ab: 28,
+          az: 52,
+          ba: 53,
+          bb: 54,
+          aaa: 703,
+          abt: 748,
+          bxf: 1982
+        };
+        Object.keys(mapping).forEach(function(key) {
+          assert.equal(
+            list._letterToNum(key), mapping[key],
+            key + ' did not properly convert to ' + mapping[key]);
+        });
+      });
+    });
   });
 
   describe('db.charms.addFromCharmData', function() {
