+++ karma-mocha-old.conf.js.tmpl
@@ -62,7 +62,6 @@ module.exports = function(config) {
       // It's possible we can remove some or all of these above.
 
       'jujugui/static/gui/src/test/test_bundle_import_notifications.js',
-      'jujugui/static/gui/src/test/test_console.js',
       'jujugui/static/gui/src/test/test_d3_components.js',
       'jujugui/static/gui/src/test/test_entity_extension.js',
       'jujugui/static/gui/src/test/test_env_change_set.js',
